+++ include/TopArtists.php
@@ -116,7 +116,7 @@
     						</div>';
     						
 							$m=0; 
-						$st=3+(60/$page_in)*$playtime/($count_max*200);    				
+						$st=(int)3+(60/(int)$page_in)*(int)$playtime/((int)$count_max*200);    				
     					echo'      	   
       	   				<div class="'; 
       	   				if($st>strlen($playtime)*3){
